// frontend/packages/identity-core/src/types/api.ts
export interface ProfileEmail {
  address: string;
  verified: boolean;
  preferred: boolean;
}